Move type defs to the correct places
